0x02-102 use long numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 void fibonacci(void)
 {
-	int i, f, prev;
+	long i, f, prev;
 
 	i = 0;
